add structure for readonly values
